requirements.txt
flask
flask-sqlalchemy
flask-cors
python-dotenv
psycopg2-binary
